updated Fire Element
FireElement now works with a "combo" system, where main ability generates 1 fire-chi, and the secondary puts the fire-chi to 2. When Main is used with 2 fire chi, it gets buffed

+++ Source/un_EZ/FireElement.h
@@ -56,9 +56,9 @@ public:
 		float fireKick = 6000;
 
 	UPROPERTY(EditAnywhere, Category = "Ability2")
-		float maxFireChi = 2;
+		int maxFireChi = 2;
 
-	float fireChi;
+	int fireChi;
 
 	UPROPERTY(EditAnywhere, Category = "Ability2")
 		FVector boostedAbility1Scale;
